Refactor out code for response time series chart from AdminHomePageComponent
