+++ src/web/app/pages-admin/admin-home-page/admin-home-page.component.html
@@ -1,9 +1,7 @@
-<div class="card bg-light top-padded">
-  <div class="line-chart">
-    <svg></svg>
-  </div>
+<div>
+  <tm-response-time-series-chart>
+  </tm-response-time-series-chart>
 </div>
-
 <div class="card bg-light top-padded">
   <div class="card-body">
     <div>
