+++ src/web/app/pages-admin/admin-home-page/admin-home-page.module.ts
@@ -4,6 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 import { AjaxLoadingModule } from '../../components/ajax-loading/ajax-loading.module';
 import { AdminHomePageComponent } from './admin-home-page.component';
+import {ResponseTimeSeriesChartModule} from "../../components/response-time-series-chart/response-time-series-chart.module";
 
 const routes: Routes = [
   {
@@ -22,11 +23,12 @@ const routes: Routes = [
   exports: [
     AdminHomePageComponent,
   ],
-  imports: [
-    CommonModule,
-    FormsModule,
-    RouterModule.forChild(routes),
-    AjaxLoadingModule,
-  ],
+    imports: [
+        CommonModule,
+        FormsModule,
+        RouterModule.forChild(routes),
+        AjaxLoadingModule,
+        ResponseTimeSeriesChartModule,
+    ],
 })
 export class AdminHomePageModule { }
